Phase 4: Fix all ClangWarn/MSVCWarn implicit conversion warnings in tests
All 29 warnings surfaced by ClangWarn (strict warnings, non-error) fixed
across 7 test/benchmark files. Library source is unaffected.

Pattern: duration.count() and size_t values used in floating-point
expressions without explicit cast. Fix: static_cast<double>() on all
integer operands in mixed arithmetic.

Specific fixes:
- test_platform_optimizations.cpp: 8 x -Wimplicit-int-float-conversion
  (scalar_time/simd_time/read_time/write_time .count() in bandwidth/speedup)
- test_cpu_detection.cpp: 4 x -Wimplicit-int-float-conversion
  (seq_time/simd_time .count() in bandwidth/speedup calculations)
- test_numerical_safety.cpp: 2 x -Wold-style-cast (double) casts,
  2 x -Wimplicit-int-float-conversion, 1 x -Wsign-conversion
  (crashes int subtracted from size_t stress_iterations)
- test_error_handling_comprehensive.cpp: 2 x -Wold-style-cast,
  2 x -Wimplicit-int-float-conversion (size_t i used in double arithmetic),
  1 x size_t iterations used as double divisor
- test_math_comprehensive.cpp: 2 x -Wimplicit-int-float-conversion
  (vector_time.count() denominator in ErfPerformance/GammaPerformance)
- test_simd_policy.cpp: 1 x -Wsign-conversion
  (int loop variable passed to shouldUseSIMD(size_t))
- benchmark_simd_all.cpp: 2 x -Wimplicit-int-float-conversion
  (size_ and results_.size() used as double divisors)

MSVCWarn build also clean (uses different warning subset).
ClangStrict test file issues remain pre-existing and tracked separately.

+++ tests/test_math_comprehensive.cpp
@@ -278,7 +278,7 @@ TEST_F(MathUtilsTest, ErfPerformance) {
     end = std::chrono::high_resolution_clock::now();
     auto vector_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
 
-    double speedup = static_cast<double>(scalar_time.count()) / vector_time.count();
+    double speedup = static_cast<double>(scalar_time.count()) / static_cast<double>(vector_time.count());
 
     // We expect at least no significant slowdown
     EXPECT_GT(speedup, 0.5) << "Vectorized should not be significantly slower than scalar";
@@ -307,7 +307,7 @@ TEST_F(MathUtilsTest, GammaPerformance) {
     end = std::chrono::high_resolution_clock::now();
     auto vector_time = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
 
-    double speedup = static_cast<double>(scalar_time.count()) / vector_time.count();
+    double speedup = static_cast<double>(scalar_time.count()) / static_cast<double>(vector_time.count());
 
     // We expect at least no significant slowdown
     EXPECT_GT(speedup, 0.5) << "Vectorized should not be significantly slower than scalar";
